Fixes composed transformers to deal with empty transformation results
When a transformation results in an empty array it should throw a
TransformationException. As composed transformers call other transformers,
these exceptions should be taken into account, and ignore them, considering
that no result was obtained.

This commit introduces this change and also updates the composed transformers
to throw an exception when the transformation result is empty.