Redesigns the application to make it pluggable
The application architecture has been redesigned and restructured so
that it is now pluggable. This means that now it is possible to add new
transformers to the GUI and/or CLI by just extending the
org.sing_group.seda.plugin.spi.SedaPluginFactory interface and adding the
extending class to a
META-INF/services/org.sing_group.seda.plugin.spi.SedaPluginFactory file.