Add support to toggle data factories
Add support to allow users to select the data factory to use when
processing a set of files: "In memory" selected indicates that the
DefaultDatatypeFactory must be used and deselected that the
LazyDatatypeFactory must be used.

Also make that all transformations accept a DatatypeFactory, using
this way of invoking them from the GUI plugins.