Add the NCBI rename plugin
This commit adds the NCBI rename plugin, which uses the NCBI web
to obtain information from accession numbers and renames sequence
files and sequence headers using this information.

This plugin relies on the
MapRenameSequencesGroupDatasetTransformation, which is a generic
way of achieving the same behaviour by providing a replacements
map. This plugin also has an option to save such replacements map
obtained from the NCBI.