Fix line breaks preservation
This commit fixes a bug (spread over several classes) that was
causing that line breaks diferent than "\n" (i.e. Windows
"\r\n") were not preserved when writing the files again. The bug
was twofold:
- By one hand, the line break string was not being saved when
the file was redaded.
- On the other hand, most of transformations would loose this
information because the transformed SequencesGroups were not
being created using the previous properties map.

To avoid this kind of errors, methods to create SequenceGroup
objects without a properties map have been removed as well.