Update manual to include Docker images usage
This commit updates the manual to include explanations about the Docker images to
run external software (Clustal Omega and BLAST). A new 'Installation and
Configuration' section is also added to the manual to explain the Docker
installation in the main OS.

This commit also updates the Dockerfile to set a property to tell SEDA to hide these
options when running inside the Docker container.