Fix a bug when loading an empty file using LazyFileSequencesGroup
This commit fixes a bug when loading an empty file using a LazyFileSequencesGroup
object. Since the file is empty, the file properties could not be determined, so
they should be ignored instead of trying to read them raising an exception.

This way, operations like Blast work properly and can produce empty files if there
are no hits.