Fix SAPP commands configuration
This commit fixes the SAPP commands configuration, which in the case of the Docker
execution cannot be made by using JFileChooserPanel GUI controls because they do not
allow to select files inside the Docker image. For this, a new class for the Docker
configuration of this commands is introduced, using JTextField controls for user
input.