Update distributable versions and installers
This commit updates the distributable versions and installers to:
- Add all the new modules to the Windows 64b and MacOS distributable versions and
installers.
- Remove all modules in the Windows 32b version, which now only consists of the core
SEDA module. This is because there are no available versions of the external software
required by these modules for Windows 32b.
- Disable the local execution of Splign/Compart, ProSplign/ProCompart, EMBOSS and
bedtools in the Windows 64b version. This is because there are no executable binaries
for this plattform and therefore they can be only executed via Docker.
- Disable the local execution of Splign/Compart and ProSplign/Procompart in the MacOS
version. This is because there are no executable binaries for this plattform and
therefore they can be only executed via Docker.