Introduce support for saving and loading operation configurations
This commit introduces support for saving and loading operation configurations. To do
so, SedaGuiPlugin defines three new methods to allow the main SEDA application to
know if the current operation configuration can be saved and to save and load it in
case it is possible.

Operation configurations are saved and loaded into/from JSON files making use of the
JsonObjectReader and JsonObjectWriter classes.

The first operation supporting this feature is the FilteringSedaGuiPlugin.