Add support for GZip file and unifies data factories
SEDA now supports reading Fasta files compressed with GZip. When using
lazy datatypes, the compressed Fasta is firstly uncompressed in a
temporal file, so that it can be treated as a regular file.

This commit also includes an important changes in the data factories. On
one hand, the main change for the default ("in memory") data factory is
that the sequences are now loaded into memory instead of reading them
from the files (they were "lazy" sequences). On the other hand, the main
change for the lazy data factory is that a sequences group will not
re-create a sequence received through the constructor that already is a
lazy sequence. This avoids the creation of repeated files for the same
sequence.

This changes in the data factories make them more uniform, as the
default now works completely in memory and the lazy now works completely
on files without duplicating them.

Finally, some refactoring, both in the main and tests code, has been
done in the classes related to the new functionalities.