Improve, optimize and refactorize file loading
File loading was too slow due to the use of the same algorithm
independently of the datatype used to access the Fasta information. This
commit improves this by using a different algorithm depending on the
target datatype and on the charset of the input file.

Charset support was also improved in several ways. First, the number of
datasets supported was restricted to a reduced list, as it could not be
guarantied that data loading worked with all the datasets supported by
Java. Second, some problems with the BOM bytes used by UTF were solved.
Finally, a different and more simple algorithm is used for charsets with
one byte by char, which greatly improves loading speed with these
datasets.

In addition, datatype classes were refactorized by moving all the
datatype classes to the same "datatype" package and by renaming the
Default* classes to InMemory* and the Lazy* classes to InDisk*.