Clear file selection to avoid user mistakes
This commit makes the neccessary changes to clear the file name
in the shared JFileChooser to avoid user mistakes. So far, the
last file name selected was maintained in subsequent openings of
the file chooser and some users reported that this behaviour was
confusing and error-prone (e.g. select first a FASTA file and
then a configuration file).