Fix bug in ProSplign/ProCompart pipeline
This commit fixes an important bug in the ProSplign/ProCompart
pipeline. The tblastn output must be sorted by subject and query
and so far it was not being sorted properly. This commit fixes
this behaviour and now the output is sorted properly by subject
and query as the procompart tool requires.