Fix bug in BLAST operations using Docker
This commit fixes a bug in BLAST operations using Docker. The
bug appeared when using BLAST commands that require protein
databases (such as blastp) and an alias was used. In those cases,
the alias file is named "*.pal" and SEDA was only looking for a
"*nal" file (which is used in nucleotide databases). As SEDA
failed to find the correct alias file, the database paths could
not be mounted into the Docker image and the command failed.