Fix Splign/Compart protocol
This commit fixes the protocol to properly retrieve the FASTA sequences from the Splign/Compart annotations created in the output.tsv file.
So fwar, We were creating a *bidirectional genome* (the original genome plus its reverse-complement) but we realized that Splign/Compart
is able to anotate in both strands and therefore such *bidirectional genome* is no longer needed. After removing this, the way that the
FASTA sequences are extracted from the annotations varies depending on the query/subject start/end coordinates. This is described in the
corresponding section of the manual.