Fixes some PARAM_*_HELP constants
This commit fixes some PARAM_*_HELP constants to make them more
informative, specially the help associated to the sequence case
option in the reformat file command.

Also, this commit starts using PARAM_*_HELP instead of
PARAM_*_DESCRIPTION in some commands. This must be applied to
all options so that we can offer the greater amount of
information to the CLI users.